+++ server/src/api/routes/names.ts
@@ -1,13 +1,13 @@
 import { Router } from 'express';
 import * as controller from 'api/controllers/names';
-import * as validator from '../modules/names/name.validator';
+import * as validations from 'api/validations/names';
 
 const api = Router();
 
-api.get('/', validator.index, controller.index);
-api.post('/', validator.submit, controller.submit);
-api.get('/:id', validator.details, controller.details);
-api.post('/:id/approve', validator.approve, controller.approve);
-api.post('/:id/deny', validator.deny, controller.deny);
+api.get('/', validations.index, controller.index);
+api.post('/', validations.submit, controller.submit);
+api.get('/:id', validations.details, controller.details);
+api.post('/:id/approve', validations.approve, controller.approve);
+api.post('/:id/deny', validations.deny, controller.deny);
 
 export default api;
